Move CommitmentProof to poly/commitment.
