+++ src/core/poly/commitment.rs
@@ -69,3 +69,14 @@ impl<F: ExtensionOf<BaseField>, H: Hasher> PolynomialDecommitment<F, H> {
             .verify(root, queries.deref().clone())
     }
 }
+
+pub struct CommitmentProof<F: ExtensionOf<BaseField>, H: Hasher> {
+    pub decommitment: PolynomialDecommitment<F, H>,
+    pub commitment: H::Hash,
+}
+
+impl<F: ExtensionOf<BaseField> + IntoSlice<H::NativeType>, H: Hasher> CommitmentProof<F, H> {
+    pub fn verify(&self, queries: &Queries) -> bool {
+        self.decommitment.verify(self.commitment, queries)
+    }
+}
